Rebase Package details
