+++ src/main/java/com/design/pattern/Design/Pattern/Observer/Design/Pattern/Observer/NotificationAlertObserver.java
@@ -1,0 +1,6 @@
+package com.design.pattern.Design.Pattern.Observer.Design.Pattern.Observer;
+
+public interface NotificationAlertObserver {
+
+    void update();
+}
