--- src/main/java/com/design/pattern/Design/Pattern/ObserverDesignPattern/Observer/NotificationAlertObserver.java
@@ -1,6 +1,0 @@
-package com.design.pattern.Design.Pattern.ObserverDesignPattern.Observer;
-
-public interface NotificationAlertObserver {
-
-    void update();
-}
