+++ src/main/java/com/design/pattern/Design/Pattern/Strategy/Design/Pattern/WithStrategyDesignPattern/Main.java
@@ -1,4 +1,4 @@
-package com.design.pattern.Design.Pattern.StrategyDesignPattern.WithStrategyDesignPattern;
+package com.design.pattern.Design.Pattern.Strategy.Design.Pattern.WithStrategyDesignPattern;
 
 public class Main {
     public static void main(String[] args) {
