+++ src/main/java/com/design/pattern/Design/Pattern/Strategy/Design/Pattern/WithStrategyDesignPattern/Stretegy/DriveStrategy.java
@@ -1,0 +1,5 @@
+package com.design.pattern.Design.Pattern.Strategy.Design.Pattern.WithStrategyDesignPattern.Stretegy;
+
+public interface DriveStrategy {
+    public void drive();
+}
