+++ src/main/java/com/design/pattern/Design/Pattern/Strategy/Design/Pattern/WithoutStrategyDesignPattern/GoodsVehicle.java
@@ -1,0 +1,5 @@
+package com.design.pattern.Design.Pattern.Strategy.Design.Pattern.WithoutStrategyDesignPattern;
+
+public class GoodsVehicle extends Vehicle {
+
+}
