+++ src/main/java/com/design/pattern/Design/Pattern/Strategy/Design/Pattern/WithoutStrategyDesignPattern/OffRoadVehicle.java
@@ -1,4 +1,4 @@
-package com.design.pattern.Design.Pattern.StrategyDesignPattern.WithoutStrategyDesignPattern;
+package com.design.pattern.Design.Pattern.Strategy.Design.Pattern.WithoutStrategyDesignPattern;
 
 public class OffRoadVehicle extends Vehicle {
