+++ src/main/java/com/design/pattern/Design/Pattern/Strategy/Design/Pattern/WithoutStrategyDesignPattern/Vehicle.java
@@ -1,4 +1,4 @@
-package com.design.pattern.Design.Pattern.StrategyDesignPattern.WithoutStrategyDesignPattern;
+package com.design.pattern.Design.Pattern.Strategy.Design.Pattern.WithoutStrategyDesignPattern;
 
 public class Vehicle {
     public void drive() {
