--- src/main/java/com/design/pattern/Design/Pattern/StrategyDesignPattern/WithStrategyDesignPattern/OffRoadVehicle.java
@@ -1,9 +1,0 @@
-package com.design.pattern.Design.Pattern.StrategyDesignPattern.WithStrategyDesignPattern;
-
-import com.design.pattern.Design.Pattern.StrategyDesignPattern.WithStrategyDesignPattern.Stretegy.SportyDrive;
-
-public class OffRoadVehicle extends Vehicle {
-    OffRoadVehicle() {
-        super(new SportyDrive());
-    }
-}
