--- src/main/java/com/design/pattern/Design/Pattern/StrategyDesignPattern/WithStrategyDesignPattern/Stretegy/DriveStrategy.java
@@ -1,5 +1,0 @@
-package com.design.pattern.Design.Pattern.StrategyDesignPattern.WithStrategyDesignPattern.Stretegy;
-
-public interface DriveStrategy {
-    public void drive();
-}
