--- src/main/java/com/design/pattern/Design/Pattern/StrategyDesignPattern/WithoutStrategyDesignPattern/GoodsVehicle.java
@@ -1,5 +1,0 @@
-package com.design.pattern.Design.Pattern.StrategyDesignPattern.WithoutStrategyDesignPattern;
-
-public class GoodsVehicle extends Vehicle {
-
-}
